+++ src/scippneutron/_utils/__init__.py
